CSCETSIN-703:Change the banner that shows whether dataset is removed/deprecated
Bug fix:  Modified the index.jsx to fetch all version datasets to the constructor via Stores so the 'deprecated' information could be fetched from the version datasets directly ('deprecated' doesn't show in version_set like 'removed').

Now the banner shows whether the dataset has been removed or deprecated, and if there is an existing version it will provide a link to it.

Should not have side-effects